Moved loading indicator into its own component and added simple tests for it
