+++ assets/src/scripts/components/loadingIndicator.js
